+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `areAllVariationsSelected` subject to `condition.product`.
+
+### Fixed
+- Prevent `condition.else` and other condition blocks to be rendered at the same time.
+
+### Changed
+- Invalid condition error will only appear in the `development` environment.
 
 ## [1.2.0] - 2020-06-02
 ### Added